+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-05-21
+
+### Changed
+
+- YouTube URLs as reference links will now display the video thumbnail and link above the build. It no longer needs to specifically be a YouTube embed URL to work.
+
+### Fixed
+
+- Fixed typo in Transpose mutator description.
+
 ## 2024-05-20
 
 ### Added